Introduced MappingGenerator.
Some functions and definitions were moved to get the program compiled.
