+++ src/mapping_metadata.h
@@ -14,6 +14,8 @@ class Index;
 class CandidateProcessor;
 class PairedEndMappingMetadata;
 template <typename MappingRecord>
+class MappingGenerator;
+template <typename MappingRecord>
 class Chromap;
 
 class MappingMetadata {
@@ -131,6 +133,8 @@ class MappingMetadata {
   friend class Index;
   friend class CandidateProcessor;
   friend class PairedEndMappingMetadata;
+  template <typename MappingRecord>
+  friend class MappingGenerator;
   template <typename MappingRecord>
   friend class Chromap;
 };
